+++ source/Tests/Infrastructure/OutgoingMessages/AggregationResult/AggregationResultDocumentWriterTests.cs
@@ -12,20 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Globalization;
 using System.IO;
-using System.Linq;
-using System.Text.Json;
 using System.Threading.Tasks;
 using Application.OutgoingMessages.Common;
-using DocumentValidation;
 using Domain.OutgoingMessages;
 using Domain.Transactions.Aggregations;
 using Infrastructure.Configuration.Serialization;
 using Infrastructure.OutgoingMessages.AggregationResult;
 using Infrastructure.OutgoingMessages.Common;
-using Json.Schema;
 using NodaTime.Text;
 using Tests.Factories;
 using Tests.Fixtures;
@@ -184,142 +179,3 @@ public class AggregationResultDocumentWriterTests : IClassFixture<DocumentValida
 }
 
 #pragma warning disable
-internal sealed class AssertAggregationResultJsonDocument : IAssertAggregationResultDocument
-{
-    private readonly JsonSchemaProvider _schemas = new(new CimJsonSchemas());
-    private readonly JsonDocument _document;
-    private readonly JsonElement _root;
-
-    public AssertAggregationResultJsonDocument(Stream document)
-    {
-        _document = JsonDocument.Parse(document);
-        _root = _document.RootElement.GetProperty("NotifyAggregatedMeasureData_MarketDocument");
-    }
-
-    public IAssertAggregationResultDocument HasMessageId(string expectedMessageId)
-    {
-        Assert.Equal(expectedMessageId, _root.GetProperty("mRID").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasSenderId(string expectedSenderId)
-    {
-        Assert.Equal(expectedSenderId, _root.GetProperty("sender_MarketParticipant.mRID").GetProperty("value").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasReceiverId(string expectedReceiverId)
-    {
-        Assert.Equal(expectedReceiverId, _root.GetProperty("receiver_MarketParticipant.mRID").GetProperty("value").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasTimestamp(string expectedTimestamp)
-    {
-        Assert.Equal(expectedTimestamp, _root.GetProperty("createdDateTime").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasTransactionId(Guid expectedTransactionId)
-    {
-        Assert.Equal(expectedTransactionId, _root.GetProperty("Series").EnumerateArray().ToList()[0].GetProperty("mRID").GetGuid());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasGridAreaCode(string expectedGridAreaCode)
-    {
-        Assert.Equal(expectedGridAreaCode, _root.GetProperty("Series").EnumerateArray().ToList()[0]
-            .GetProperty("meteringGridArea_Domain.mRID")
-            .GetProperty("value")
-            .ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasBalanceResponsibleNumber(string expectedBalanceResponsibleNumber)
-    {
-        Assert.Equal(expectedBalanceResponsibleNumber, _root.GetProperty("Series").EnumerateArray().ToList()[0]
-            .GetProperty("balanceResponsibleParty_MarketParticipant.mRID")
-            .GetProperty("value").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasEnergySupplierNumber(string expectedEnergySupplierNumber)
-    {
-        Assert.Equal(expectedEnergySupplierNumber, _root.GetProperty("Series").EnumerateArray().ToList()[0]
-            .GetProperty("energySupplier_MarketParticipant.mRID")
-            .GetProperty("value").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasProductCode(string expectedProductCode)
-    {
-        Assert.Equal(expectedProductCode, _root.GetProperty("Series").EnumerateArray().ToList()[0].GetProperty("product").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasPeriod(string expectedStartOfPeriod, string expectedEndOfPeriod)
-    {
-        Assert.Equal(expectedStartOfPeriod, _root.GetProperty("Series").EnumerateArray().ToList()[0]
-            .GetProperty("Period")
-            .GetProperty("timeInterval")
-            .GetProperty("start")
-            .GetProperty("value").ToString());
-        Assert.Equal(expectedEndOfPeriod, _root.GetProperty("Series").EnumerateArray().ToList()[0]
-            .GetProperty("Period")
-            .GetProperty("timeInterval")
-            .GetProperty("end")
-            .GetProperty("value").ToString());
-        return this;
-    }
-
-    public IAssertAggregationResultDocument HasPoint(int position, int quantity)
-    {
-        var point = _root.GetProperty("Series").EnumerateArray().ToList()[0]
-            .GetProperty("Period")
-            .GetProperty("Point").EnumerateArray().ToList()[position - 1];
-
-        Assert.NotNull(point);
-        Assert.Equal(position, point.GetProperty("position").GetProperty("value").GetInt32());
-        Assert.Equal(quantity, point.GetProperty("quantity").GetInt32());
-        return this;
-    }
-
-    public async Task<IAssertAggregationResultDocument> DocumentIsValidAsync()
-    {
-        var schema = await _schemas.GetSchemaAsync<JsonSchema>("NOTIFYAGGREGATEDMEASUREDATA", "0").ConfigureAwait(false);
-        var validationOptions = new EvaluationOptions()
-        {
-            OutputFormat = OutputFormat.List,
-        };
-        var validationResult = schema.Evaluate(_document, validationOptions);
-        var errors = validationResult.Details.Where(detail => detail.HasErrors).Select(x => x.Errors).ToList()
-            .SelectMany(e => e.Values).ToList();
-        Assert.True(validationResult.IsValid, string.Join("\n", errors));
-        return this;
-    }
-
-    public IAssertAggregationResultDocument SettlementMethodIsNotPresent()
-    {
-        throw new NotImplementedException();
-    }
-
-    public IAssertAggregationResultDocument EnergySupplierNumberIsNotPresent()
-    {
-        throw new NotImplementedException();
-    }
-
-    public IAssertAggregationResultDocument BalanceResponsibleNumberIsNotPresent()
-    {
-        throw new NotImplementedException();
-    }
-
-    public IAssertAggregationResultDocument QuantityIsNotPresentForPosition(int position)
-    {
-        throw new NotImplementedException();
-    }
-
-    public IAssertAggregationResultDocument QualityIsNotPresentForPosition(int position)
-    {
-        throw new NotImplementedException();
-    }
-}
